changed time_created to int

+++ src/http_tests/channel_messages_http_test.py
@@ -70,8 +70,8 @@ def test_channel_id_not_valid(url, initialise_user_data):
     check that channel_messages raises InputError
     if Channel ID is not a valid channel
     '''
-    user1_credentials = initialise_user_data['user1']      
-    invalid_channel_id = -1 
+    user1_credentials = initialise_user_data['user1']
+    invalid_channel_id = -1
 
     response = requests.get(f"{url}/channel/messages", params={
         'token': user1_credentials['token'],
@@ -122,7 +122,7 @@ def test_return_type(url, initialise_user_data, initialise_channel_data):
     assert isinstance(message_payload['messages'][0]['message_id'], int)
     assert isinstance(message_payload['messages'][0]['u_id'], int)
     assert isinstance(message_payload['messages'][0]['message'], str)
-    assert isinstance(message_payload['messages'][0]['time_created'], float)
+    assert isinstance(message_payload['messages'][0]['time_created'], int)
 
     assert isinstance(message_payload['start'], int)
     assert isinstance(message_payload['end'], int)
